+++ src/components/MaterialIcon.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import clsx from 'clsx'
+
+export default function MaterialIcon({
+  className,
+  children,
+  ...props
+}: React.PropsWithChildren<
+  { className?: string } & JSX.IntrinsicElements['i']
+>) {
+  return (
+    <i className={clsx('material-icons', className)} aria-hidden {...props}>
+      {children}
+    </i>
+  )
+}