+++ src/components/renderer/FormElementLabelContainer.tsx
@@ -4,6 +4,7 @@ import clsx from 'clsx'
 import { FormTypes } from '@oneblink/types'
 import useReplaceableText from '../../hooks/useReplaceableText'
 import QuillHTML from '../QuillHTML'
+import MaterialIcon from '../MaterialIcon'
 
 function FormElementLabelContainer({
   className,
@@ -64,7 +65,7 @@ export function HintTooltip({
       leaveTouchDelay={10000}
       id={`${inputId}-hint`}
     >
-      <i className="material-icons has-text-grey ob-label__hint">info</i>
+      <MaterialIcon className="has-text-grey ob-label__hint">info</MaterialIcon>
     </Tooltip>
   )
 }