+++ src/form-elements/FormElementTelephone.tsx
@@ -7,6 +7,7 @@ import FormElementLabelContainer from '../components/renderer/FormElementLabelCo
 import { FormElementValueChangeHandler, IsDirtyProps } from '../types/form'
 import { LookupNotificationContext } from '../hooks/useLookupNotification'
 import useElementAriaDescribedby from '../hooks/useElementAriaDescribedby'
+import MaterialIcon from '../components/MaterialIcon'
 
 type Props = {
   id: string
@@ -62,7 +63,7 @@ function FormElementTelephone({
               aria-describedby={ariaDescribedby}
             />
             <span className="ob-input-icon icon is-small is-right">
-              <i className="material-icons is-size-5">phone</i>
+              <MaterialIcon className="is-size-5">phone</MaterialIcon>
             </span>
           </div>
           {!!element.readOnly && !!value && (